refactor: Image 패키지 삭제, File 패키지로 통합
