+++ src/main/java/com/macrosoft/modakserver/domain/file/controller/FileController.java
@@ -1,6 +1,6 @@
 package com.macrosoft.modakserver.domain.file.controller;
 
-import com.macrosoft.modakserver.domain.file.service.FileService;
+import com.macrosoft.modakserver.domain.file.service.FileServiceImpl;
 import com.macrosoft.modakserver.global.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/files")
 @Tag(name = "File API", description = "S3로 파일를 업로드 하기 위해 사용되는 API 입니다.")
 public class FileController {
-    private final FileService fileService;
+    private final FileServiceImpl fileServiceImpl;
 
     @Operation(summary = "미리 서명된 URL 받기", description = "S3 버켓에 파일을 업로드하기 위한 미리 서명된 URL을 받습니다.")
     @GetMapping("/presigned-url/{extension}")
@@ -25,6 +25,6 @@ public class FileController {
             @Parameter(description = "파일 확장자", example = "webp")
             @PathVariable(name = "extension")
             String extension) {
-        return BaseResponse.onSuccess(fileService.getPresignedUrl(extension));
+        return BaseResponse.onSuccess(fileServiceImpl.getPresignedUrl(extension));
     }
 }
