--- src/main/java/com/macrosoft/modakserver/domain/image/service/ImageService.java
@@ -1,5 +1,0 @@
-package com.macrosoft.modakserver.domain.image.service;
-
-public interface ImageService {
-    void deleteImageFromS3(String imageUrl);
-}
